Revert @mysten/sui changes; copy reservation helpers into wallet-sdk
The address-balance work doesn't need to expose anything new from the
sui sdk. Restore `packages/sui/src/utils/{coin-reservation,index}.ts`
to their pre-branch state, drop the @mysten/sui changeset, and add a
local `packages/wallet-sdk/src/transaction-analyzer/coin-reservation.ts`
with the helpers the analyzer consumes.

While here:

- Drop the `payFromAddressBalance` flag + its withdrawFromCoin branch.
  With `payment: []` tx.gas can't be used in the PTB, so a zero-balance
  tracked coin is enough — any draw from it flags as an over-split.
- `TrackedCoin.consume` now nulls `ownerAddress` alongside balance, so
  a later write into a consumed coin drops silently instead of being
  credited back to the original holder at end-of-PTB.
- Trim verbose inline comments.
