+++ packages/sui/src/utils/index.ts
@@ -38,10 +38,3 @@ export { deriveDynamicFieldID } from './dynamic-fields.js';
 
 export { deriveObjectID } from './derived-objects.js';
 export { normalizeTypeTag } from '../bcs/type-tag-serializer.js';
-
-export {
-	isCoinReservationDigest,
-	parseAccumulatorFieldCoinType,
-	parseCoinReservationBalance,
-	xorCoinReservationObjectId,
-} from './coin-reservation.js';
