+++ packages/wallet-sdk/test/transaction-analyzer/coin-flows-intents.test.ts
@@ -591,10 +591,13 @@ describe('Coin Flows - input reservation refs', () => {
 		expect(results.coins.issues!.some((i) => /could not be resolved/.test(i.message))).toBe(true);
 	});
 
-	it('splitCoins(tx.gas, ...) with empty payment charges AB without flagging over-split', async () => {
+	it('splitCoins(tx.gas, ...) with empty payment flags over-split (tx.gas unusable)', async () => {
 		const client = new MockSuiClient();
+		// Use an address that has no default SUI coins so the mock's plugin
+		// leaves `payment: []` in place rather than falling back to a coin.
+		const abOnlySender = normalizeSuiAddress('0xab000001');
 		const tx = new Transaction();
-		tx.setSender(DEFAULT_SENDER);
+		tx.setSender(abOnlySender);
 		const [split] = tx.splitCoins(tx.gas, [2_000_000_000n]);
 		tx.transferObjects([split], tx.pure.address('0x456'));
 
@@ -603,10 +606,11 @@ describe('Coin Flows - input reservation refs', () => {
 		json.gasData.budget = '10000000';
 
 		const results = await analyze({ balanceFlows }, { client, transaction: JSON.stringify(json) });
-		expect(results.balanceFlows.issues).toBeUndefined();
-		const senderSui = results.balanceFlows.result?.sender.find((f) => f.coinType === SUI);
-		// 2_000_000_000 split out + 10_000_000 gas budget, both charged to AB
-		expect(senderSui?.amount).toBe(-2_010_000_000n);
+		// With payment:[] tx.gas can't be used in the PTB. The analyzer flags
+		// any draw from it as over-split against a zero-balance coin.
+		expect(results.balanceFlows.issues?.some((i) => /SplitCoins at command/.test(i.message))).toBe(
+			true,
+		);
 	});
 
 	it('emits an issue when the unmasked object is not a balance accumulator field', async () => {
